Add files via upload
Horiseon is a website that helps business owners find feasible clients.It controls the online opinion that their clients will have access to. It determines the platforms that are most compatible with their brand.As a developer, I've been tasked with refactoring Horiseon's website's code so that it meets accessibility standards. To accomplish this, I've made changes to the HTML and CSS files.